Projects/theinsightedge/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.2.1
click==8.3.1
email-validator==2.1.0
fastapi==0.135.1
h11==0.16.0
idna==3.11
passlib==1.7.4
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.1
python-jose==3.3.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.36
